added common-model and refactoring to use it (#488)
* move schema out of submission

* sperate migration and schema mangers

* move ClinicalEntitySchemaNames to common model

* move donor entity extractors to common-model

* rename schemaManager to dictionaryManager

* move enums to remove circular dependency

* update swagger for convenience

* append migration to dry-run and run endpoints

* update swagger for clarity

* fix swagger endpoints
