--- src/lectern-client/index.ts
@@ -1,5 +1,0 @@
-import * as schemaManager from '../submission/schema/schema-manager';
-
-export default {
-  schemaManager,
-};
