+++ src/resources/swagger.yaml
@@ -55,10 +55,10 @@ paths:
         '200':
           description: 'Sample registration and Clinical Submission disabled state has been set'
 
-  /submission/schema:
+  /dictionary:
     get:
       tags:
-        - Schemas
+        - Dictionary
       responses:
         '200':
           description: 'the current loaded schema by the clinical submission'
@@ -68,9 +68,11 @@ paths:
                 type: array
                 items:
                   type: object
-    patch:
+  /dictionary/migration/run:
+    post:
       tags:
-        - Schemas
+        - Dictionary Migration
+        - Admin
       description: 'update and load a new schema version from lectern to be used.'
       requestBody:
         content:
@@ -88,10 +90,11 @@ paths:
             application/json:
               schema:
                 type: object
-  /submission/schema/dry-run-update:
+  /dictionary/migration/dry-run-update:
     post:
       tags:
-        - Schemas
+        - Dictionary Migration
+        - Admin
       requestBody:
         content:
           application/json:
@@ -108,10 +111,11 @@ paths:
             application/json:
               schema:
                 type: object
-  /submission/schema/migration/resume:
+  /dictionary/migration/resume:
     post:
       tags:
-        - Schemas
+        - Dictionary Migration
+        - Admin
       responses:
         '200':
           description: 'Resume an active migration'
@@ -121,10 +125,11 @@ paths:
                 type: array
                 items:
                   type: object
-  /submission/schema/migration/{migrationId}:
+  /dictionary/migration/{migrationId}:
     get:
       tags:
-        - Schemas
+        - Dictionary Migration
+        - Admin
       parameters:
         - name: migrationId
           description: the id of a specific migration to fetch
@@ -141,10 +146,11 @@ paths:
                 type: array
                 items:
                   type: object
-  /submission/schema/migration/:
+  /dictionary/migration/:
     get:
       tags:
-        - Schemas
+        - Dictionary Migration
+        - Admin
       responses:
         '200':
           description: 'A list of all migrations'
@@ -154,10 +160,10 @@ paths:
                 type: array
                 items:
                   type: object
-  /submission/schema/changes:
+  /dictionary/changes:
     get:
       tags:
-        - Schemas
+        - Dictionary
       parameters:
         - name: from
           description: the old version
@@ -178,10 +184,10 @@ paths:
             application/json:
               schema:
                 type: object
-  /submission/schema/list:
+  /dictionary/list:
     get:
       tags:
-        - Schemas
+        - Dictionary
       parameters:
         - name: includeFields
           description: include fields of each schema in the response
@@ -199,10 +205,10 @@ paths:
                 items:
                   type: object
 
-  /submission/schema/template/all:
+  /dictionary/template/all:
     get:
       tags:
-        - Schemas
+        - Dictionary
       responses:
         '200':
           description: 'A zip file with all schemas'
@@ -211,7 +217,7 @@ paths:
               schema:
                 type: string
                 format: binary
-  /submission/schema/template/{schemaName}:
+  /dictionary/template/{schemaName}:
     parameters:
       - name: schemaName
         description: the name of the schema to generate the template for
@@ -222,7 +228,7 @@ paths:
           type: string
     get:
       tags:
-        - Schemas
+        - Dictionary
       responses:
         '200':
           description: 'A tsv template file for the given schemaName'
@@ -614,6 +620,7 @@ paths:
     post:
       tags:
         - ICGC Import
+        - Admin
       summary: transform icgc donor samples data to argo format sample registration
       responses:
         '200':
@@ -639,6 +646,7 @@ paths:
     post:
       tags:
         - ICGC Import
+        - Admin
       summary: imports the json body (donors) to the clinical donors collection obtained from the preprocess api
       description:
       requestBody:
@@ -752,6 +760,7 @@ paths:
     patch:
       tags:
         - Patch Operations
+        - Admin
       summary: Used to override donor aggregatedInfoStats
       description: For coreCompletionOverride, only the ones defined will be overridden the rest will be recalculated
       parameters:
@@ -785,6 +794,66 @@ paths:
         '200':
           description: 'Donor with updated stats'
 
+  /submission/schema:
+    get:
+      deprecated: true
+      summary: 'use  `GET /dictionary` instead'
+      tags:
+        - Deprecated
+    patch:
+      deprecated: true
+      summary: 'use  `POST /dictionary/migration/run` instead'
+      tags:
+        - Deprecated
+  /submission/schema/dry-run-update:
+    post:
+      deprecated: true
+      summary: 'use `POST /dictionary/migration/dry-run-update` instead'
+      tags:
+        - Deprecated
+  /submission/schema/migration/resume:
+    post:
+      deprecated: true
+      summary: 'use `POST /dictionary/migration/resume` instead'
+      tags:
+        - Deprecated
+  /submission/schema/migration/{migrationId}:
+    get:
+      deprecated: true
+      summary: 'use `GET /dictionary/migration/{migrationId}` instead'
+      tags:
+        - Deprecated
+  /submission/schema/migration/:
+    get:
+      deprecated: true
+      summary: 'use `GET /dictionary/migration` instead'
+      tags:
+        - Deprecated
+  /submission/schema/changes:
+    get:
+      deprecated: true
+      summary: 'use `GET /dictionary/changes` instead'
+      tags:
+        - Deprecated
+  /submission/schema/list:
+    get:
+      deprecated: true
+      summary: 'use `GET /dictionary/list` instead'
+      tags:
+        - Deprecated
+  /submission/schema/template/all:
+    get:
+      deprecated: true
+      summary: 'use `GET /dictionary/template/all` instead'
+      tags:
+        - Deprecated
+  /submission/schema/template/{schemaName}:
+    get:
+      deprecated: true
+      summary: 'use `GET /dictionary/template/{schemaName}` instead'
+      tags:
+        - Deprecated
+
 servers:
   - url: /
 tags:
@@ -792,8 +861,10 @@ tags:
     description: Service status monitoring
   - name: Submission Configurations
     description: Submission Service Configurations
-  - name: Schemas
-    description: View and manage the Data Schema used by Argo-Clinical
+  - name: Dictionary
+    description: View the Data Dictionary used by Argo-Clinical
+  - name: Dictionary Migration
+    description: Update Data Dictionary and manage migrations
   - name: Sample Registration
     description: Registering new Donors, Specimens, and Samples
   - name: Submission
@@ -804,6 +875,10 @@ tags:
     description: Clinical Service Patch Operations
   - name: ICGC Import
     description: APIs to do the icgc import
+  - name: Admin
+    description: Admin operations collected for convenience
+  - name: Deprecated
+    description: Deprecated enpoints
 components:
   parameters:
     QueryProgramId:
