+++ src/submission/validation-clinical/therapy.ts
@@ -1,17 +1,18 @@
 import {
   SubmissionValidationError,
-  ClinicalEntitySchemaNames,
-  TreatmentFieldsEnum,
   SubmittedClinicalRecord,
   DataValidationErrors,
-  ClinicalTherapyType,
-  TherapyRxNormFields,
 } from '../submission-entities';
+import {
+  ClinicalEntitySchemaNames,
+  TreatmentFieldsEnum,
+  ClinicalTherapyType,
+} from '../../common-model/entities';
 import { DeepReadonly } from 'deep-freeze';
 import { Donor, Treatment } from '../../clinical/clinical-entities';
 import * as utils from './utils';
 import _ from 'lodash';
-import { getSingleClinicalObjectFromDonor } from '../submission-to-clinical/submission-to-clinical';
+import { getSingleClinicalObjectFromDonor } from '../../common-model/functions';
 
 export const validate = async (
   therapyRecord: DeepReadonly<SubmittedClinicalRecord>,
