+++ src/submission/validation-clinical/treatment.ts
@@ -1,18 +1,20 @@
 import {
   SubmissionValidationError,
-  ClinicalEntitySchemaNames,
-  TreatmentFieldsEnum,
   SubmittedClinicalRecord,
   DataValidationErrors,
+} from '../submission-entities';
+import {
+  ClinicalEntitySchemaNames,
   ClinicalUniqueIdentifier,
+  TreatmentFieldsEnum,
   ClinicalTherapyType,
   ClinicalTherapySchemaNames,
-} from '../submission-entities';
+} from '../../common-model/entities';
 import { DeepReadonly } from 'deep-freeze';
 import { Donor, Treatment } from '../../clinical/clinical-entities';
 import * as utils from './utils';
 import _ from 'lodash';
-import { getSingleClinicalObjectFromDonor } from '../submission-to-clinical/submission-to-clinical';
+import { getSingleClinicalObjectFromDonor } from '../../common-model/functions';
 import { checkClinicalEntityDoesntBelongToOtherDonor } from './utils';
 
 export const validate = async (
